[#305] Added abstract models.
`models` module replaced by a `models` package containing an `abstract`
module (for abstract models) and a `generic` module (for the default
models, previously in the `models` module)

---

Resolves celery/django-celery-results#305

Fixes celery/django-celery-results#314
