+++ django_celery_results/models/__init__.py
@@ -1,0 +1,3 @@
+from .generic import ChordCounter, GroupResult, TaskResult
+
+__ALL__ = [ChordCounter, GroupResult, TaskResult]
